Refactor rc/mod.rs tests
