+++ crates/objc2/src/rc/mod.rs
@@ -74,58 +74,3 @@ pub use self::id_traits::{DefaultId, SliceId, SliceIdMut};
 pub use self::ownership::{Owned, Ownership, Shared};
 pub use self::test_object::{__RcTestObject, __ThreadTestData};
 pub use self::weak_id::WeakId;
-
-#[cfg(test)]
-mod tests {
-    use core::marker::PhantomData;
-    use core::mem::size_of;
-
-    use super::{Id, Owned, Ownership, Shared, WeakId};
-    use crate::runtime::Object;
-
-    #[repr(C)]
-    struct TestType {
-        inner: Object,
-    }
-
-    #[repr(C)]
-    struct MyObject<'a> {
-        inner: Object,
-        p: PhantomData<&'a str>,
-    }
-
-    /// Test that `Id<T, O>` is covariant over `T`.
-    #[allow(unused)]
-    fn assert_id_variance<'a, 'b, O: Ownership>(
-        obj: &'a Id<MyObject<'static>, O>,
-    ) -> &'a Id<MyObject<'b>, O> {
-        obj
-    }
-
-    /// Test that `WeakId<T>` is covariant over `T`.
-    #[allow(unused)]
-    fn assert_weak_id_variance<'a, 'b>(
-        obj: &'a WeakId<MyObject<'static>>,
-    ) -> &'a WeakId<MyObject<'b>> {
-        obj
-    }
-
-    #[test]
-    fn test_size_of() {
-        assert_eq!(size_of::<Id<TestType, Owned>>(), size_of::<&TestType>());
-        assert_eq!(size_of::<Id<TestType, Shared>>(), size_of::<&TestType>());
-        assert_eq!(
-            size_of::<Option<Id<TestType, Owned>>>(),
-            size_of::<&TestType>()
-        );
-        assert_eq!(
-            size_of::<Option<Id<TestType, Shared>>>(),
-            size_of::<&TestType>()
-        );
-
-        assert_eq!(
-            size_of::<Option<WeakId<TestType>>>(),
-            size_of::<*const ()>()
-        );
-    }
-}
